Updated urls and changed name of default function components
