+++ frontend/src/app/admin/dashboard/page.tsx
@@ -9,27 +9,27 @@ export default function Dashboard() {
       <div className={styles.gridContainer}>
         <DashboardCard
           imageURI="/dashboard_eventcreator.png"
-          url="/eventcreator"
+          url="/event-creator"
           title="Event Creator"
           last_updated="Month XX, XXXX, XX:XX"
         />
         <DashboardCard
           imageURI="/dashboard_pageeditor.png"
-          url="/pageeditor"
+          url="/page-editor"
           title="Page Editor"
           last_updated="Month XX, XXXX, XX:XX"
         />
 
         <DashboardCard
           imageURI="/dashboard_newsletter.png"
-          url="/newslettercreator"
+          url="/newsletter-creator"
           title="Newsletter"
           last_updated="Month XX, XXXX, XX:XX"
         />
 
         <DashboardCard
           imageURI="/dashboard_mailinglist.png"
-          url="/mailinglist"
+          url="/mailing-list"
           title="Mailing List"
           last_updated="Month XX, XXXX, XX:XX"
         />
