+++ frontend/src/app/admin/page-editor/home/page.tsx
@@ -12,22 +12,27 @@ import PageToggle from "@/components/PageToggle";
 
 // import PageEditorCard from "@/components/PageEditorCard";
 
-export default function Dashboard() {
+export default function HomeEditor() {
   const [isEdited, setIsEdited] = useState(false);
   const [phSubtitle, setPhSubtitle] = useState<string>("");
   const [s1Subtitle, setS1Subtitle] = useState<string>("");
   const [s1Text, setS1Text] = useState<string>("");
+  const [s2Subtitle, setS2Subtitle] = useState<string>("");
+  const [s2Text, setS2Text] = useState<string>("");
 
   /* Get page data from MongoDB */
   let pageText;
   useEffect(() => {
-    getPageText("Newsletter")
+    getPageText("Home")
       .then((response) => {
         if (response.success) {
           pageText = response.data;
           setPhSubtitle(pageText.pageSections[0].subtitle ?? "");
           setS1Subtitle(pageText.pageSections[1].sectionTitle ?? "");
           setS1Text(pageText.pageSections[1].sectionSubtitle ?? "");
+          setS2Subtitle(pageText.pageSections[2].sectionTitle ?? "");
+          setS2Text(pageText.pageSections[2].sectionSubtitle ?? "");
+          console.log("response.data: ", response.data);
         } else {
           alert(response.error);
         }
@@ -44,8 +49,12 @@ export default function Dashboard() {
       setPhSubtitle(event.target.value);
     } else if (event.target.id === "Section 1: Section Title") {
       setS1Subtitle(event.target.value);
-    } else if (event.target.id === "Section 1: Section Subtitle") {
+    } else if (event.target.id === "Section 1: Body Text") {
       setS1Text(event.target.value);
+    } else if (event.target.id === "Section 2: Section Title") {
+      setS2Subtitle(event.target.value);
+    } else if (event.target.id === "Section 2: Body Text") {
+      setS2Text(event.target.value);
     }
   };
 
@@ -55,7 +64,7 @@ export default function Dashboard() {
       console.log("Save changes");
       updatePage({
         //Pass edited text to MongoDB
-        page: "Newsletter",
+        page: "Home",
         pageSections: [
           {
             subtitle: phSubtitle,
@@ -63,6 +72,10 @@ export default function Dashboard() {
           {
             sectionTitle: s1Subtitle,
             sectionSubtitle: s1Text,
+          },
+          {
+            sectionTitle: s2Subtitle,
+            sectionSubtitle: s2Text,
           },
         ],
       })
@@ -84,13 +97,15 @@ export default function Dashboard() {
     // Implement cancel logic
     if (isEdited) {
       console.log("Cancel changes");
-      getPageText("Newsletter")
+      getPageText("Home")
         .then((response) => {
           if (response.success) {
             pageText = response.data;
             setPhSubtitle(pageText.pageSections[0].subtitle ?? "");
             setS1Subtitle(pageText.pageSections[1].sectionTitle ?? "");
             setS1Text(pageText.pageSections[1].sectionSubtitle ?? "");
+            setS2Subtitle(pageText.pageSections[2].sectionTitle ?? "");
+            setS2Text(pageText.pageSections[2].sectionSubtitle ?? "");
           } else {
             alert(response.error);
           }
@@ -104,26 +119,26 @@ export default function Dashboard() {
 
   return (
     <main className={styles.page}>
-      <PageToggle
-        pages={["Our Impact", "Testimonials", "Newsletter"]}
-        links={["./impact", "./testimonials", "./newsletter"]}
-        currPage={2}
-        refreshPage={true}
-      />
+      <PageToggle pages={["Home"]} links={["./home"]} currPage={0} />
       <div className={styles.sectionContainer}>
         <Collapsable
           title="Page Header"
-          subsection={["Subtitle"]}
+          subsection={["Subtitle", "Header Image Carousel"]}
           textbox={[phSubtitle]}
           onChange={handleEdit}
         />
         <Collapsable
           title="Section 1"
-          subsection={["Section Title", "Section Subtitle"]}
+          subsection={["Section Title", "Body Text"]}
           textbox={[s1Subtitle, s1Text]}
           onChange={handleEdit}
         />
-
+        <Collapsable
+          title="Section 2"
+          subsection={["Section Title", "Body Text", "Sponsor Image Gallery"]}
+          textbox={[s2Subtitle, s2Text]}
+          onChange={handleEdit}
+        />
         <div className={styles.buttonContainer}>
           <CancelButton
             text="Cancel"
