+++ frontend/src/app/admin/page-editor/impact/page.tsx
@@ -5,40 +5,31 @@ import { getPageText, updatePage } from "../../../../api/pageeditor";
 
 import styles from "./page.module.css";
 
-import AlertBanner from "@/components/AlertBanner";
 import Button from "@/components/Button";
 import CancelButton from "@/components/CancelButton";
 import Collapsable from "@/components/Collapsable";
 import PageToggle from "@/components/PageToggle";
+// import { PhpSharp } from "@mui/icons-material";
 
 // import PageEditorCard from "@/components/PageEditorCard";
 
-export default function Dashboard() {
+export default function ImpactEditor() {
   const [isEdited, setIsEdited] = useState(false);
+
   const [phSubtitle, setPhSubtitle] = useState<string>("");
   const [s1Subtitle, setS1Subtitle] = useState<string>("");
-  const [s1Text, setS1Text] = useState<string>("");
   const [s2Subtitle, setS2Subtitle] = useState<string>("");
-  const [s2Text, setS2Text] = useState<string>("");
-  const [s3Subtitle, setS3Subtitle] = useState<string>("");
-  const [s3Text, setS3Text] = useState<string>("");
-
-  const [showAlert, setShowAlert] = useState(false);
 
   /* Get page data from MongoDB */
   let pageText;
   useEffect(() => {
-    getPageText("About Us")
+    getPageText("Our Impact")
       .then((response) => {
         if (response.success) {
           pageText = response.data;
           setPhSubtitle(pageText.pageSections[0].subtitle ?? "");
           setS1Subtitle(pageText.pageSections[1].sectionTitle ?? "");
-          setS1Text(pageText.pageSections[1].sectionSubtitle ?? "");
           setS2Subtitle(pageText.pageSections[2].sectionTitle ?? "");
-          setS2Text(pageText.pageSections[2].sectionSubtitle ?? "");
-          setS3Subtitle(pageText.pageSections[3].sectionTitle ?? "");
-          setS3Text(pageText.pageSections[3].sectionSubtitle ?? "");
           console.log("response.data: ", response.data);
         } else {
           alert(response.error);
@@ -52,20 +43,12 @@ export default function Dashboard() {
   /* Handle Fields upon edit */
   const handleEdit = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
     setIsEdited(true);
-    if (event.target.id === "Page Subtitle: Subtitle") {
+    if (event.target.id === "Page Header: Subtitle") {
       setPhSubtitle(event.target.value);
-    } else if (event.target.id === "Section 1 - Our Mission: Section Title") {
+    } else if (event.target.id === "Section 1 - Testimonials: Subtitle") {
       setS1Subtitle(event.target.value);
-    } else if (event.target.id === "Section 1 - Our Mission: Body Text") {
-      setS1Text(event.target.value);
-    } else if (event.target.id === "Section 2 - Our Team: Section Title") {
+    } else if (event.target.id === "Section 2 - Newsletter: Subtitle") {
       setS2Subtitle(event.target.value);
-    } else if (event.target.id === "Section 2 - Our Team: Body Text") {
-      setS2Text(event.target.value);
-    } else if (event.target.id === "Section 3 - Contact Us: Section Title") {
-      setS3Subtitle(event.target.value);
-    } else if (event.target.id === "Section 3 - Contact Us: Body Text") {
-      setS3Text(event.target.value);
     }
   };
 
@@ -75,22 +58,16 @@ export default function Dashboard() {
       console.log("Save changes");
       updatePage({
         //Pass edited text to MongoDB
-        page: "About Us",
+        page: "Our Impact",
         pageSections: [
           {
             subtitle: phSubtitle,
           },
           {
             sectionTitle: s1Subtitle,
-            sectionSubtitle: s1Text,
           },
           {
             sectionTitle: s2Subtitle,
-            sectionSubtitle: s2Text,
-          },
-          {
-            sectionTitle: s3Subtitle,
-            sectionSubtitle: s3Text,
           },
         ],
       })
@@ -112,17 +89,13 @@ export default function Dashboard() {
     // Implement cancel logic
     if (isEdited) {
       console.log("Cancel changes");
-      getPageText("About Us")
+      getPageText("Our Impact")
         .then((response) => {
           if (response.success) {
             pageText = response.data;
             setPhSubtitle(pageText.pageSections[0].subtitle ?? "");
             setS1Subtitle(pageText.pageSections[1].sectionTitle ?? "");
-            setS1Text(pageText.pageSections[1].sectionSubtitle ?? "");
             setS2Subtitle(pageText.pageSections[2].sectionTitle ?? "");
-            setS2Text(pageText.pageSections[2].sectionSubtitle ?? "");
-            setS3Subtitle(pageText.pageSections[3].sectionTitle ?? "");
-            setS3Text(pageText.pageSections[3].sectionSubtitle ?? "");
           } else {
             alert(response.error);
           }
@@ -134,46 +107,34 @@ export default function Dashboard() {
     }
   };
 
-  const handleCloseAlert = () => {
-    setShowAlert(false);
-  };
-
   return (
     <main className={styles.page}>
-      <div className={styles.alert}>
-        {showAlert && <AlertBanner text={"Event Details Saved!"} onClose={handleCloseAlert} />}
-      </div>
       <PageToggle
-        pages={["About Us", "Our Mission", "Our Team", "Contact Us"]}
-        links={["./about", "./mission", "./team", "./contact"]}
+        pages={["Our Impact", "Testimonials", "Newsletter"]}
+        links={["./impact", "./testimonials", "./newsletter"]}
         currPage={0}
         refreshPage={true}
       />
       <div className={styles.sectionContainer}>
         <Collapsable
-          title="Page Subtitle"
+          title="Page Header"
           subsection={["Subtitle"]}
-          textbox={[phSubtitle, ""]}
+          textbox={[phSubtitle]}
           onChange={handleEdit}
         />
         <Collapsable
-          title="Section 1 - Our Mission"
-          subsection={["Section Title", "Body Text", "Section Image"]}
-          textbox={[s1Subtitle, s1Text]}
+          title="Section 1 - Testimonials"
+          subsection={["Subtitle"]}
+          textbox={[s1Subtitle]}
           onChange={handleEdit}
         />
         <Collapsable
-          title="Section 2 - Our Team"
-          subsection={["Section Title", "Body Text", "Section Image"]}
-          textbox={[s2Subtitle, s2Text]}
+          title="Section 2 - Newsletter"
+          subsection={["Subtitle"]}
+          textbox={[s2Subtitle]}
           onChange={handleEdit}
         />
-        <Collapsable
-          title="Section 3 - Contact Us"
-          subsection={["Section Title", "Body Text", "Section Image"]}
-          textbox={[s3Subtitle, s3Text]}
-          onChange={handleEdit}
-        />
+
         <div className={styles.buttonContainer}>
           <CancelButton
             text="Cancel"
