+++ frontend/src/app/admin/page-editor/newsletter/page.tsx
@@ -9,28 +9,25 @@ import Button from "@/components/Button";
 import CancelButton from "@/components/CancelButton";
 import Collapsable from "@/components/Collapsable";
 import PageToggle from "@/components/PageToggle";
-// import { PhpSharp } from "@mui/icons-material";
 
 // import PageEditorCard from "@/components/PageEditorCard";
 
-export default function Dashboard() {
+export default function NewsletterEditor() {
   const [isEdited, setIsEdited] = useState(false);
-
   const [phSubtitle, setPhSubtitle] = useState<string>("");
   const [s1Subtitle, setS1Subtitle] = useState<string>("");
-  const [s2Subtitle, setS2Subtitle] = useState<string>("");
+  const [s1Text, setS1Text] = useState<string>("");
 
   /* Get page data from MongoDB */
   let pageText;
   useEffect(() => {
-    getPageText("Our Impact")
+    getPageText("Newsletter")
       .then((response) => {
         if (response.success) {
           pageText = response.data;
           setPhSubtitle(pageText.pageSections[0].subtitle ?? "");
           setS1Subtitle(pageText.pageSections[1].sectionTitle ?? "");
-          setS2Subtitle(pageText.pageSections[2].sectionTitle ?? "");
-          console.log("response.data: ", response.data);
+          setS1Text(pageText.pageSections[1].sectionSubtitle ?? "");
         } else {
           alert(response.error);
         }
@@ -45,10 +42,10 @@ export default function Dashboard() {
     setIsEdited(true);
     if (event.target.id === "Page Header: Subtitle") {
       setPhSubtitle(event.target.value);
-    } else if (event.target.id === "Section 1 - Testimonials: Subtitle") {
+    } else if (event.target.id === "Section 1: Section Title") {
       setS1Subtitle(event.target.value);
-    } else if (event.target.id === "Section 2 - Newsletter: Subtitle") {
-      setS2Subtitle(event.target.value);
+    } else if (event.target.id === "Section 1: Section Subtitle") {
+      setS1Text(event.target.value);
     }
   };
 
@@ -58,16 +55,14 @@ export default function Dashboard() {
       console.log("Save changes");
       updatePage({
         //Pass edited text to MongoDB
-        page: "Our Impact",
+        page: "Newsletter",
         pageSections: [
           {
             subtitle: phSubtitle,
           },
           {
             sectionTitle: s1Subtitle,
-          },
-          {
-            sectionTitle: s2Subtitle,
+            sectionSubtitle: s1Text,
           },
         ],
       })
@@ -89,13 +84,13 @@ export default function Dashboard() {
     // Implement cancel logic
     if (isEdited) {
       console.log("Cancel changes");
-      getPageText("Our Impact")
+      getPageText("Newsletter")
         .then((response) => {
           if (response.success) {
             pageText = response.data;
             setPhSubtitle(pageText.pageSections[0].subtitle ?? "");
             setS1Subtitle(pageText.pageSections[1].sectionTitle ?? "");
-            setS2Subtitle(pageText.pageSections[2].sectionTitle ?? "");
+            setS1Text(pageText.pageSections[1].sectionSubtitle ?? "");
           } else {
             alert(response.error);
           }
@@ -112,7 +107,7 @@ export default function Dashboard() {
       <PageToggle
         pages={["Our Impact", "Testimonials", "Newsletter"]}
         links={["./impact", "./testimonials", "./newsletter"]}
-        currPage={0}
+        currPage={2}
         refreshPage={true}
       />
       <div className={styles.sectionContainer}>
@@ -123,15 +118,9 @@ export default function Dashboard() {
           onChange={handleEdit}
         />
         <Collapsable
-          title="Section 1 - Testimonials"
-          subsection={["Subtitle"]}
-          textbox={[s1Subtitle]}
-          onChange={handleEdit}
-        />
-        <Collapsable
-          title="Section 2 - Newsletter"
-          subsection={["Subtitle"]}
-          textbox={[s2Subtitle]}
+          title="Section 1"
+          subsection={["Section Title", "Section Subtitle"]}
+          textbox={[s1Subtitle, s1Text]}
           onChange={handleEdit}
         />
